remove duplicate zip_directory()
